requirements.txt
alabaster==0.7.16
yt_dlp
altair==5.2.0
attrs==23.2.0
audioread==3.0.1
Babel==2.14.0
beautifulsoup4==4.12.3
black==24.2.0
blinker==1.7.0
build==1.0.3
CacheControl==0.13.1
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
crashtest==0.4.1
cycler==0.12.1
decorator==5.1.1
distlib==0.3.8
docutils==0.20.1
dulwich==0.21.7
entrypoints==0.4
Faker==23.2.1
fastjsonschema==2.19.1
favicon==0.7.0
filelock==3.13.1
fonttools==4.49.0
fsspec==2024.2.0
gitdb==4.0.11
GitPython==3.1.42
htbuilder==0.6.2
idna==3.6
imagesize==1.4.1
importlib-metadata==7.0.1
installer==0.7.0
jaraco.classes==3.3.1
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
keyring==24.3.0
kiwisolver==1.4.5
lazy_loader==0.3
librosa==0.10.1
llvmlite==0.42.0
lxml==5.1.0
Markdown==3.5.2
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.5
matplotlib==3.8.3
mdurl==0.1.2
more-itertools==10.2.0
mpmath==1.3.0
msgpack==1.0.7
mypy-extensions==1.0.0
networkx==3.2.1
numba==0.59.0
numpy==1.26.4
openai-whisper==20231117
packaging==23.2
pandas==2.2.0
pathspec==0.12.1
pexpect==4.9.0
pillow==10.2.0
pkginfo==1.9.6
platformdirs==4.2.0
pooch==1.8.1
prometheus_client==0.20.0
protobuf==4.25.3
ptyprocess==0.7.0
pyarrow==15.0.0
pycparser==2.21
pydeck==0.8.0
pydub==0.25.1
Pygments==2.17.2
pymdown-extensions==10.7
pyparsing==3.1.1
pyproject_hooks==1.0.0
python-dateutil==2.8.2
pytube==15.0.0
pytz==2024.1
pywin32-ctypes==0.2.2
PyYAML==6.0.1
rapidfuzz==3.6.1
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.0
rpds-py==0.18.0
scikit-learn==1.4.1.post1
scipy==1.12.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
snowballstemmer==2.2.0
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.7
st-annotated-text==4.0.1
streamlit==1.31.1
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.0
streamlit-embedcode==0.1.2
streamlit-extras==0.4.0
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.6
streamlit-keyup==0.2.3
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.3.0
tiktoken==0.6.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.3
toolz==0.12.1
torch==2.2.0
tornado==6.4
tqdm==4.66.2
trove-classifiers==2024.1.31
typing_extensions==4.9.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
validators==0.22.0
virtualenv==20.25.0
watchdog==4.0.0
wget==3.2
zipp==3.17.0
